Add auditing functionality to User entity
New fields (createdBy, createdAt, modifiedBy, modifiedAt) are added to the User entity to track changes. Additionally, JPA's auditing functionality has been enabled and a new class (AuditorAwareImplements) is implemented to provide the current auditor. This will help in maintaining a clear record of who created or modified an entity and when it happened.