Refactor profile update method in AccountService
The name of the method responsible for updating user profiles in the AccountService and AccountController has been changed from "updateUserProfile" to "updateProfile" for simplification. Additionally, an 'updateEmail' function has been added in IAccountService interface and its implementation in AccountService. Tests have been adjusted accordingly to reflect the method name changes.