Add feature to update user status and refactor messages
A new feature has been implemented to allow the status of a user in an admin role to be updated. Specifically, the statuses of 'activate' and 'deactivate' have been made available. Additionally, the structure of messages has been refactored for more generic use, using 'ENTITY' instead of specific 'USER' and 'ROLE'. This results in more reusable and manageable code.