Add test for enum messages and update message format
The `MessagesTest` class has been implemented to add tests for asserting the formatting of different error and informational messages in the system. Additionally, the formatting for the messages in the `Messages` enum has also been updated to be more consistent and clear, particularly around the naming of various entities in the system.