Update test cases and adjust application configurations
The commit includes a variety of updates to several test cases, making adjustments to assertions and renames certain entities in test cases for better readability. Changes also include shifting the timestamps used in ExceptionHandler tests to be epoch days instead of the current date. Additionally, this commit turns off SQL display and formatting in JPA configuration for testing.